pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Kurt Wolf", email = "kurtwolfbuilds@gmail.com"},
]
dependencies = [
    "django>=4.1.4",
    "djangorestframework>=3.14.0",
    "pydantic>=1.10.2",
    "pyyaml>=6.0",
    "uritemplate>=4.1.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
